requirements.txt
python-docx~=1.1.2
tkinterdnd2~=0.3.0